common: better type conversions (#82)
* More type conversion stuff.

simple_reorder has many changes where the conversion to data_t<type_o>
is done a little too early.  one case of moving the alpha,beta test
up a few lines in ker lambda ker (to mirror other cases).
Also a couple of cases where intermediate calcs are forced to
float/double, and then finally converted to some data_t.
(some of these may need to be revisited for int data_t support, later,
 but at least the conversions are visible in the code for now.)

* Update cpu_simple_concat.hpp

comment about int conversion safety.
change loop variable type to match int num_arrs.

* Update simple_reorder.hpp

a few more cases of delaying conversion to data_t<type_o> from upstream impls

* Update gemm_convolution.hpp

indentation

* Update cpu_simple_sum.hpp

deal with more sxc++ int vs size_t warnings

* Update ref_eltwise.cpp

sxc++ mod.  elu_bwd delays cast to final type.  Could also consider casting alpha down to float, so double->float warning doesn't happen.

* gtests eltwise: harmonize and ref calcs in double
* common: fix conversions in c++ api

+++ include/mkldnn.hpp
@@ -385,9 +385,9 @@ struct memory: public primitive  {
                 format aformat) {
             validate_dims(adims);
             error::wrap_c_api(
-                    mkldnn_memory_desc_init(&data, adims.size(),
+                    mkldnn_memory_desc_init(&data, (int)adims.size(),
                         adims.size() == 0 ? nullptr : &adims[0],
-                         convert_to_c(adata_type), convert_to_c(aformat)),
+                        convert_to_c(adata_type), convert_to_c(aformat)),
                     "could not initialize a memory descriptor");
         }
 
@@ -697,7 +697,7 @@ struct concat : public primitive {
             auto c_api_inputs = cpp_to_c(inputs);
 
             error::wrap_c_api(mkldnn_concat_primitive_desc_create(
-                    &result, &output.data, c_api_inputs.size(),
+                    &result, &output.data, (int)c_api_inputs.size(),
                     concat_dimension, &c_api_inputs[0]),
                 "could not create a concat primitive descriptor");
             reset(result);
@@ -710,8 +710,8 @@ struct concat : public primitive {
             auto c_api_inputs = cpp_to_c(inputs);
 
             error::wrap_c_api(mkldnn_concat_primitive_desc_create(
-                    &result, nullptr, c_api_inputs.size(), concat_dimension,
-                    &c_api_inputs[0]),
+                    &result, nullptr, (int)c_api_inputs.size(),
+                    concat_dimension, &c_api_inputs[0]),
                 "could not create a concat primitive descriptor");
             reset(result);
         }
@@ -766,7 +766,7 @@ struct sum : public primitive {
             auto c_api_inputs = cpp_to_c(inputs);
 
             error::wrap_c_api(mkldnn_sum_primitive_desc_create(
-                    &result, &output.data, c_api_inputs.size(),
+                    &result, &output.data, (int)c_api_inputs.size(),
                     &scale[0], &c_api_inputs[0]),
                 "could not create a sum primitive descriptor");
             reset(result);
@@ -779,7 +779,7 @@ struct sum : public primitive {
             auto c_api_inputs = cpp_to_c(inputs);
 
             error::wrap_c_api(mkldnn_sum_primitive_desc_create(
-                    &result, nullptr, c_api_inputs.size(), &scale[0],
+                    &result, nullptr, (int)c_api_inputs.size(), &scale[0],
                     &c_api_inputs[0]),
                 "could not create a sum primitive descriptor");
             reset(result);
